Renamed the package name back to org.bjoss.arquillian.spring.integration.test.annotation.
