+++ arquillian-service-integration-spring-javaconfig/src/test/java/org/jboss/arquillian/spring/integration/javaconfig/model/PackagesAnnotatedClass.java
@@ -17,7 +17,7 @@
 
 package org.jboss.arquillian.spring.integration.javaconfig.model;
 
-import org.jboss.arquillian.spring.integration.annotation.test.SpringAnnotationConfiguration;
+import org.jboss.arquillian.spring.integration.test.annotation.SpringAnnotationConfiguration;
 
 /**
  * <p>Simple class used for testing the enricher.</p>
